Fix webapp version selector bug
