+++ common/src/main/java/com/toastedbits/plugins/codeconnect/neo4j/CodeConnectNeo4jUpstreamReporter.java
@@ -1,9 +1,8 @@
-package com.toastedbits.plugins.codeconnect;
+package com.toastedbits.plugins.codeconnect.neo4j;
 
 import java.io.OutputStream;
 import java.io.PrintWriter;
 import java.util.ArrayList;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -14,6 +13,9 @@ import javax.ws.rs.client.Entity;
 import javax.ws.rs.client.WebTarget;
 import javax.ws.rs.core.Response;
 
+import com.toastedbits.plugins.codeconnect.CodeConnectUpstreamReporter;
+import com.toastedbits.plugins.codeconnect.DependencyMapping;
+import com.toastedbits.plugins.codeconnect.MavenCoordinate;
 import com.toastedbits.plugins.codeconnect.entities.Neo4jTransactionPostEntity;
 import com.toastedbits.plugins.codeconnect.entities.Neo4jTransactionStatement;
 import com.toastedbits.plugins.codeconnect.exceptions.CodeConnectException;
@@ -49,7 +51,7 @@ public class CodeConnectNeo4jUpstreamReporter implements CodeConnectUpstreamRepo
 	
 			int transactionId = neo4j.beginTransaction(client);
 			resetNode(mapping.getProject(), transactionId);
-			reportDependencies(transactionId);
+			reportDependencies(mapping.getProject(), transactionId);
 			neo4j.commitTransaction(client, transactionId);
 		}
 		finally {
@@ -57,53 +59,13 @@ public class CodeConnectNeo4jUpstreamReporter implements CodeConnectUpstreamRepo
 		}
 	}
 
-	private Map<String, String> buildNodeParameters(MavenCoordinate coordinate) {
-		Map<String, String> parameters = new LinkedHashMap<>();
-		parameters.put("projectName", coordinate.getArtifactId());
-		parameters.put("projectGroup", coordinate.getGroupId());
-		parameters.put("projectVersion", coordinate.getVersion());
-		return parameters;
-	}
-
-	private Neo4jTransactionStatement buildMergeStatement(MavenCoordinate coordinate) {
-		Neo4jTransactionStatement statement = new Neo4jTransactionStatement();
-		statement.setStatement("MERGE (proj:Project {name:{projectName}, group:{projectGroup}, version:{projectVersion}})");
-		statement.setParameters(buildNodeParameters(coordinate));
-		return statement;
-	}
-
-	private Neo4jTransactionStatement buildDeleteOutboundStatement() {
-		Neo4jTransactionStatement statement = new Neo4jTransactionStatement();
-		statement.setStatement("MATCH (down:Project {name:{projectName}, group:{projectGroup}, version:{projectVersion}})-[rel]->() DELETE rel");
-		statement.setParameters(buildNodeParameters(mapping.getProject()));
-		return statement;
-	}
-
-	private Neo4jTransactionStatement buildAssociationStatement(MavenCoordinate child, String configName, MavenCoordinate parent) {
-		Neo4jTransactionStatement statement = new Neo4jTransactionStatement();
-		//At time of writing, neo4j restApi does not support parameterization on relation name
-		statement.setStatement(
-				"MATCH (c:Project {name:{childName}, group:{childGroup}, version:{childVersion}}), " +
-				"(p:Project {name:{parentName}, group:{parentGroup}, version:{parentVersion}}) " +
-				"MERGE (c)-[:" + configName + "]->(p)");
-		Map<String, String> parameters = new LinkedHashMap<>();
-		parameters.put("childName", child.getArtifactId());
-		parameters.put("childGroup", child.getGroupId());
-		parameters.put("childVersion", child.getVersion());
-		parameters.put("parentName", parent.getArtifactId());
-		parameters.put("parentGroup", parent.getGroupId());
-		parameters.put("parentVersion", parent.getVersion());
-		statement.setParameters(parameters);
-		return statement;
-	}
-
-	private void resetNode(MavenCoordinate coordinate, int transactionId) throws CodeConnectException {
+	public void resetNode(MavenCoordinate coordinate, int transactionId) throws CodeConnectException {
 		writer.println("> Refreshing project node for: " + 
 				coordinate.getGroupId() + ":" + coordinate.getArtifactId()  + ":" + coordinate.getVersion());
 		Neo4jTransactionPostEntity payload = new Neo4jTransactionPostEntity();
 		List<Neo4jTransactionStatement> statements = new ArrayList<>();
-		statements.add(buildMergeStatement(coordinate));
-		statements.add(buildDeleteOutboundStatement());
+		statements.add(neo4j.buildMergeStatement(coordinate));
+		statements.add(neo4j.buildDeleteOutboundStatement(mapping.getProject()));
 		payload.setStatements(statements);
 
 		WebTarget target = client.target(baseUrl).path(Neo4jEndpoint.DB.Data.Transaction.path(transactionId));
@@ -114,7 +76,7 @@ public class CodeConnectNeo4jUpstreamReporter implements CodeConnectUpstreamRepo
 		}
 	}
 
-	private void reportDependencies(int transactionId) throws CodeConnectException {
+	public void reportDependencies(MavenCoordinate project, int transactionId) throws CodeConnectException {
 		for(Map.Entry<String, Set<MavenCoordinate>> configuration : mapping.getEntries()) {
 			String configName = configuration.getKey();
 			Set<MavenCoordinate> dependencies = configuration.getValue();
@@ -126,8 +88,8 @@ public class CodeConnectNeo4jUpstreamReporter implements CodeConnectUpstreamRepo
 			for(MavenCoordinate dependency : dependencies) {
 				writer.println("> (" + configName + ") " + dependency.getGroupId() + ":" + dependency.getArtifactId() + ":" + dependency.getVersion());
 				
-				statements.add(buildMergeStatement(dependency));
-				statements.add(buildAssociationStatement(mapping.getProject(), configName, dependency));
+				statements.add(neo4j.buildMergeStatement(dependency));
+				statements.add(neo4j.buildAssociationStatement(project, configName, dependency));
 			}
 
 			WebTarget target = client.target(baseUrl).path(Neo4jEndpoint.DB.Data.Transaction.path(transactionId));
