+++ gradle-plugin/src/main/java/com/toastedbits/plugins/codeconnect/ShowDownstreamTask.java
@@ -9,6 +9,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.toastedbits.plugins.codeconnect.exceptions.CodeConnectException;
+import com.toastedbits.plugins.codeconnect.neo4j.CodeConnectNeo4jDownstreamReporter;
 import com.toastedbits.plugins.codeconnect.utils.CodeConnectConfig;
 import com.toastedbits.plugins.codeconnect.utils.JerseyClientFactory;
